Edit an Delivery

+++ src/main/java/com/EAD/EAD_monolithic/repo/DeliveryRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DeliveryRepo extends JpaRepository<Delivery, Long> {
+public interface DeliveryRepo extends JpaRepository<Delivery, Integer> {
 
 }
